Match types: fix wildcard captures handling to match intent.
As demonstrated by https://github.com/lampepfl/dotty/issues/19607,
picking `H` does *not* in fact always work. It was always the
intent of the spec to pick an instantiation that would make
subtyping work, but the particular choice was wrong.

IMO this is therefore clearly a spec bug, and we can address it
irrespective of compatibility concerns.